Update SpatialAction to track the target object if it moves
Adds new attribute to GdPAIInteractable to define the max distance the
target is allowed to drift from it's planning-time location before the
action is considered to have failed.

The `max_interaction_distance` was also used to trigger the navigation
target update. This was chosen since it's the max distance an agent
can interact, so theoretically, if the target moves this distance,
we'll no longer be able to reach it if we were aiming for the center.